Moved ImageMenu component to a separate file
